requirements.txt
azure-functions==1.2.1
azure-keyvault==4.1.0
azure-loganalytics==0.1.0
azure-mgmt-compute==12.1.0
azure-mgmt-storage==4.0.0
signalfx==1.1.7
azure-identity==1.3.1